fix: restore known-good fixture symbol processing pipeline
