+++ gui/CMakeLists.txt
@@ -23,7 +23,6 @@ target_sources(${PROJECT_NAME} PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/tools/fixture_geometry_bounds.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tools/fixture_category_assignment_tool.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tools/scene_model_symbol_capture_service.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/tools/scene_model_symbol_processing.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tools/scoped_single_model_capture_scene.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tools/symbol_physical_calibration.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tools/symbol_physical_calibration_core.cpp
