+++ gui/services/fixture_symbol_processing_worker.h
@@ -34,7 +34,6 @@ public:
 
   bool Submit(FixtureSymbolProcessingRequest request);
   std::optional<FixtureSymbolProcessingResult> TakeResult();
-  bool WaitUntilIdle();
 
 private:
   // Keep the managed jthread path as the default; this fallback exists only
