+++ gui/tools/scene_model_symbol_capture_service.cpp
@@ -20,6 +20,8 @@
 namespace tools {
 namespace {
 
+constexpr float kSymbolRdpEpsilon = 1.0f;
+
 // Temporarily overrides the isolated fixture color used for extraction.
 class ScopedFixtureCaptureColor {
 public:
@@ -285,6 +287,32 @@ SceneModelSymbolRenderResult CaptureSceneModelOrthographicRenders(
     result.renders.push_back(std::move(render));
   }
   result.ok = result.renders.size() == requests.size();
+  return result;
+}
+
+// Vectorizes, simplifies, and validates a complete four-view capture set.
+SceneModelSymbolCaptureResult ProcessSceneModelOrthographicRenders(
+    std::vector<symbols::RenderedSymbolImage> renders,
+    const FixtureGeometryBounds &fixtureBoundsMm,
+    symbols::FixtureSymbolTimings *timings) {
+  SceneModelSymbolCaptureResult result;
+  result.fixtureBoundsMm = fixtureBoundsMm;
+  const auto &requests = symbols::FixtureSymbolCapturePlan();
+  if (renders.size() != requests.size()) {
+    result.error = "Could not generate all symbols from captured views.";
+    return result;
+  }
+  {
+    symbols::ScopedFixtureSymbolPhase phase(
+        timings, symbols::FixtureSymbolPhase::Vectorization);
+    result.symbols =
+        symbols::Symbol2DImageBuilder::BuildFromRenderedImages(renders);
+    for (auto &symbol : result.symbols)
+      symbols::SimplifySymbolGeometry(symbol, kSymbolRdpEpsilon);
+  }
+  result.ok = result.symbols.size() == requests.size();
+  if (!result.ok)
+    result.error = "Could not generate all symbols from captured views.";
   return result;
 }
